requirements.txt
attrs==21.4.0
iniconfig==1.1.1
packaging==21.3
pluggy==1.0.0
py==1.11.0
pyparsing==3.0.7
pytest==7.1.1
toml==0.10.2
PyTorch-lightning==1.5.10
tensorflow-cpu==2.8.0
fusepy==3.0.1
requests==2.28.1
